Move kctf-pow wrappers into a dedicated pow module
