+++ pwn/toplevel.py
@@ -37,6 +37,7 @@ from pwnlib.flag import *
 from pwnlib.fmtstr import FmtStr, fmtstr_payload, fmtstr_split
 from pwnlib.log import getLogger
 from pwnlib.memleak import MemLeak, RelativeMemLeak
+from pwnlib.pow import *
 from pwnlib.regsort import *
 from pwnlib.replacements import *
 from pwnlib.rop import ROP
